F01: Add implementation of CommuneService
