+++ src/main/java/com/carto/sn/service/CommuneService.java
@@ -1,5 +1,12 @@
 package com.carto.sn.service;
 
-public interface CommuneService {
+import java.util.List;
 
+import com.carto.sn.entities.Commune;
+
+public interface CommuneService {
+	public List<Commune> findCommuneByNomDepartement(String nomDepartement);
+	public void supprimerCommune(Long idCommune);
+	public Commune findByNomCommune(String nomCommune);
+	public List<Commune> toutesLesCommunes();
 }
